1995 Undercover TEU Caprice
This will be used for speed traps and pursuits by TEU's top 2. These will be the only two in the fleet for now. Each will be equipped with a light, sirens, cage, and a LT1 350 c.i. giving 260 horsepower. They may be used by Top 2 for regular patrol if TEU's on-duty slots are filled. Thank you.